Extract score storage into ScoreKeeper
