+++ lib/game.js
@@ -2,6 +2,7 @@ var Dinosaur = require('../lib/dinosaur');
 var Windup = require('../lib/windup');
 var GamePlay = require('../lib/game_play');
 var Levels = require('../lib/levels');
+var ScoreKeeper = require('../lib/score-keeper');
 
 function Game(canvas){
   this.canvas = canvas;
@@ -15,8 +16,8 @@ function Game(canvas){
 
 Game.prototype.play = function(){
   setKeyBindings(this);
-  loadHighScores();
-  loadHighScores2P();
+  ScoreKeeper.loadHighScores();
+  ScoreKeeper.loadHighScores2P();
   setStartScreen(gameLoop, gameLoop2P, this);
   setEndScreen(gameLoop, this);
   set2PEndScreens(gameLoop2P, this);
@@ -44,8 +45,8 @@ function gameLoop(){
   GamePlay.drawScore(game.dino, game.context);
   GamePlay.decrementFruitValues(game.fruits);
   if (GamePlay.gameOver(game.dino, game.bubbles, game.windups, game.fruits)) {
-    recordScore(game);
-    loadHighScores();
+    ScoreKeeper.recordScore(game);
+    ScoreKeeper.loadHighScores();
     GamePlay.endGameSequence(game.dino);
     return true;
   }
@@ -80,11 +81,11 @@ function gameLoop2P(){
   GamePlay.decrementFruitValues(game.fruits);
   if (GamePlay.gameOver2P(game.dino, game.dino2)) {
     if (game.dino.points > game.dino2.points){
-      recordScore2P(game, game.dino);
+      ScoreKeeper.recordScore2P(game, game.dino);
     } else {
-      recordScore2P(game, game.dino2);
+      ScoreKeeper.recordScore2P(game, game.dino2);
     }
-    loadHighScores2P();
+    ScoreKeeper.loadHighScores2P();
     clearInterval(game.intID);
     GamePlay.endGameSequence2P(game.dino, game.dino2);
     return true;
@@ -188,96 +189,5 @@ function set2PEndScreens(gameLoop2P, game) {
   }
 }
 
-function recordScore(game){
-  var scores = localStorage.getItem('high-scores');
-  if (scores) {
-    insertScore(scores, game.dino.points);
-  } else {
-    localStorage.setItem('high-scores', game.dino.points);
-  }
-}
 
-function recordScore2P(game, winner){
-  var scores = localStorage.getItem('high-scores-2p');
-  if (scores) {
-    insertScore2P(scores, winner.points);
-  } else {
-    localStorage.setItem('high-scores-2p', winner.points);
-  }
-}
-
-
-function insertScore(scores, score) {
-  var scoresArr = scores.split(" ");
-  for(var i=0; i < scoresArr.length; i++) {
-    if (score > scoresArr[i]) {
-      scoresArr.splice(i, 0, score);
-      break;
-    }
-  }
-  if (score <= scoresArr[scoresArr.length-1]) {scoresArr.push(score);}
-  localStorage.setItem('high-scores', scoresArr.slice(0, 10).join(" "));
-}
-
-function insertScore2P(scores, score) {
-  var scoresArr = scores.split(" ");
-  for(var i=0; i < scoresArr.length; i++) {
-    if (score > scoresArr[i]) {
-      scoresArr.splice(i, 0, score);
-      break;
-    }
-  }
-  if (score <= scoresArr[scoresArr.length-1]) {scoresArr.push(score);}
-  localStorage.setItem('high-scores-2p', scoresArr.slice(0, 10).join(" "));
-}
-
-function loadHighScores() {
-  removeAllNodes();
-  addHighScores();
-}
-
-function loadHighScores2P() {
-  removeAllNodes2P();
-  addHighScores2P();
-}
-
-function removeAllNodes(){
-  var highScoreList = document.getElementById("high-score-list");
-  while (highScoreList.firstChild) {
-    highScoreList.removeChild(highScoreList.firstChild);
-  }
-}
-
-function removeAllNodes2P(){
-  var highScoreList = document.getElementById("high-score-list-double");
-  while (highScoreList.firstChild) {
-    highScoreList.removeChild(highScoreList.firstChild);
-  }
-}
-
-function addHighScores(){
-  var highScoreList = document.getElementById("high-score-list");
-  var scores = localStorage.getItem('high-scores');
-  if (scores) {
-    scores = scores.split(" ");
-    scores.forEach(function(score){
-      var scoreElement = document.createElement("li");
-      scoreElement.innerHTML = score;
-      highScoreList.appendChild(scoreElement);
-    });
-  }
-}
-
-function addHighScores2P(){
-  var highScoreList = document.getElementById("high-score-list-double");
-  var scores = localStorage.getItem('high-scores-2p');
-  if (scores) {
-    scores = scores.split(" ");
-    scores.forEach(function(score){
-      var scoreElement = document.createElement("li");
-      scoreElement.innerHTML = score;
-      highScoreList.appendChild(scoreElement);
-    });
-  }
-}
 module.exports = Game;
